fix: linter error

+++ tests/e2e/dao/suite.go
@@ -203,7 +203,7 @@ func (s *E2ETestSuite) TestReissuance() {
 		// 1:  block 26: sending the reissuance result broadcast tx succeeded
 		// 2:  block 27: confirmation
 		wait = 2
-		if latestHeight%s.reissuanceEpochs == int64(wait) {
+		if latestHeight%s.reissuanceEpochs == wait {
 			break
 		}
 	}
